--- a/uber/site_sections/budget.py
+++ b/uber/site_sections/budget.py
@@ -3,13 +3,15 @@ from uber.common import *
 
 def prereg_money(session):
     preregs = defaultdict(int)
-    for attendee in session.query(Attendee).all():
+    for attendee in session.query(Attendee):
         preregs['Attendee'] += attendee.amount_paid
         preregs['extra'] += attendee.amount_extra
 
-    preregs['group_badges'] = sum(g.badge_cost for g in session.query(Group).filter(Group.tables == 0, Group.amount_paid > 0).all())
+    preregs['group_badges'] = sum(g.badge_cost for g in session.query(Group)
+                                                               .filter(Group.tables == 0, Group.amount_paid > 0)
+                                                               .options(joinedload(Group.attendees)))
 
-    dealers = session.query(Group).filter(Group.tables > 0, Group.amount_paid > 0).all()
+    dealers = session.query(Group).filter(Group.tables > 0, Group.amount_paid > 0).options(joinedload(Group.attendees)).all()
     preregs['dealer_tables'] = sum(d.table_cost for d in dealers)
     preregs['dealer_badges'] = sum(d.badge_cost for d in dealers)
 
@@ -36,11 +38,10 @@ class Root:
             'sales':   sales
         }
 
-    # TODO: add joinedload options here for efficiency
     @log_pageview
     def mpoints(self, session):
         groups = defaultdict(list)
-        for mpu in session.query(MPointsForCash).all():
+        for mpu in session.query(MPointsForCash).options(joinedload(MPointsForCash.attendee).subqueryload(Attendee.group)):
             groups[mpu.attendee and mpu.attendee.group].append(mpu)
         all = [(sum(mpu.amount for mpu in mpus), group, mpus)
                for group, mpus in groups.items()]