--- a/uber/site_sections/emails.py
+++ b/uber/site_sections/emails.py
@@ -25,10 +25,7 @@ class Root:
         count = 0
         examples = []
         email = AutomatedEmail.instances[subject]
-        attendees, groups = session.everyone()
-        models = {Attendee: attendees, Group: groups}
-        models.update({model: lister(session) for model, lister in AutomatedEmail.extra_models.items()})
-        for x in models[email.model]:
+        for x in AutomatedEmail.queries[email.model](session):
             if email.filter(x):
                 count += 1
                 url = {