… uber/templates/jobs/everywhere.html
… uber/templates/jobs/job_renderer.html
… uber/templates/jobs/signups.html
… uber/templates/registration/recent.html
… uber/templates/summary/affiliates.html
… uber/templates/summary/index.html
… uber/templates/summary/shirt_counts.html