Here are the stats on page load times before and after this round of optimizations.  As you can see, there's still a lot more slowness than we'd probably like, but it's way better, and all of the common pages especially have been much improved.

Here are the before numbers (all pages sorted from fastest to slowest):

```
0.0008997917175292969 seconds: /preregistration/kiosk
0.0010788440704345703 seconds: /preregistration/paid_preregistrations
0.002763032913208008 seconds: /registration/new
0.003251314163208008 seconds: /magprime_dept_checklist/index
0.004004478454589844 seconds: /accounts/insert_test_admin
0.009475231170654297 seconds: /common/index
0.018666744232177734 seconds: /tabletop_checkins/badged_attendees
0.02135157585144043 seconds: /signups/jobs
0.029811859130859375 seconds: /signups/login
0.033190250396728516 seconds: /preregistration/index
0.03638148307800293 seconds: /accounts/reset
0.03664684295654297 seconds: /preregistration/check_prereg
0.03684377670288086 seconds: /signups/shifts
0.03879356384277344 seconds: /tabletop_tournaments/index
0.039220333099365234 seconds: /registration/arbitrary_charge_form
0.03997159004211426 seconds: /barcode/index
0.04161477088928223 seconds: /registration/new_reg_station
0.04185152053833008 seconds: /tabletop_checkins/index
0.042601823806762695 seconds: /registration/merch
0.04282116889953613 seconds: /registration/attendee_upload
0.04623699188232422 seconds: /accounts/homepage
0.046465158462524414 seconds: /preregistration/credit_card_retry
0.04669308662414551 seconds: /tournament_admin/index
0.04710984230041504 seconds: /signups/shirt_size
0.04830765724182129 seconds: /devtools/index
0.050820350646972656 seconds: /emails/pending
0.05090928077697754 seconds: /mivs_judging/index
0.051665544509887695 seconds: /dept_checklist/index
0.05324840545654297 seconds: /magprime_dept_checklist/tech_requirements
0.05485939979553223 seconds: /accounts/change_password
0.056938886642456055 seconds: /signups/index
0.05979347229003906 seconds: /magprime_dept_checklist/treasury
0.06338644027709961 seconds: /schedule/xml
0.06380677223205566 seconds: /summary/extra_merch
0.06381702423095703 seconds: /registration/inactive
0.0655975341796875 seconds: /signups/printable
0.06892275810241699 seconds: /preregistration/stats
0.06901431083679199 seconds: /budget/mpoints
0.07081341743469238 seconds: /registration/review
0.08227992057800293 seconds: /schedule/schedule_tsv
0.0843045711517334 seconds: /registration/placeholders
0.08527207374572754 seconds: /preregistration/dealer_registration
0.0931100845336914 seconds: /static_views/magfest_js
0.09438776969909668 seconds: /emails/emails_by_interest
0.09492683410644531 seconds: /registration/discount
0.0955357551574707 seconds: /graphs/index
0.10156559944152832 seconds: /schedule/edit
0.10257816314697266 seconds: /season_supporters/index
0.11178374290466309 seconds: /schedule/panelists_owed_refunds
0.18859171867370605 seconds: /schedule/now
0.2063612937927246 seconds: /emails/index
0.2187047004699707 seconds: /hotel/index
0.2873973846435547 seconds: /devtools/gitinfo
0.29348325729370117 seconds: /registration/index
0.36557435989379883 seconds: /summary/affiliates
0.3770923614501953 seconds: /registration/feed
0.4588322639465332 seconds: /mivs_admin/index
0.5151107311248779 seconds: /registration/comments
0.517735481262207 seconds: /jobs/index
0.5530788898468018 seconds: /panel_app_management/index
0.5833683013916016 seconds: /summary/setup_teardown_neglect
0.6972801685333252 seconds: /schedule/panels_json
0.8164355754852295 seconds: /accounts/sitemap
0.8779189586639404 seconds: /dept_checklist/overview
0.903254508972168 seconds: /jobs/everywhere
0.9105856418609619 seconds: /accounts/index
1.1157405376434326 seconds: /jobs/signups
1.1459524631500244 seconds: /schedule/index
1.1655902862548828 seconds: /summary/found_how
1.194089412689209 seconds: /registration/staffers
1.2039086818695068 seconds: /groups/index
1.2924087047576904 seconds: /jobs/staffers
1.3164947032928467 seconds: /hotel/requests
1.3980891704559326 seconds: /jobs/add_volunteers_by_dept
1.4773492813110352 seconds: /summary/departments
2.033682107925415 seconds: /hotel_assignments/index
2.273275852203369 seconds: /summary/food_restrictions
2.426645517349243 seconds: /band_admin/index
2.7661099433898926 seconds: /hotel_summary/setup_teardown
2.785127878189087 seconds: /summary/shirt_counts
2.904197931289673 seconds: /summary/ratings
3.301191568374634 seconds: /hotel/hours
4.0498046875 seconds: /registration/recent
4.078104019165039 seconds: /budget/index
5.1261866092681885 seconds: /summary/food_eligible
5.28834342956543 seconds: /hotel/no_shows
7.42536997795105 seconds: /jobs/all_shifts
10.460880756378174 seconds: /summary/index
11.531702280044556 seconds: /summary/restricted_untaken
12.947009801864624 seconds: /summary/staffing_overview
13.71165943145752 seconds: /jobs/summary
16.894471883773804 seconds: /summary/consecutive_threshold
19.031463861465454 seconds: /summary/all_schedules
23.013450860977173 seconds: /summary/volunteers_owed_refunds
```

And here are the numbers after the optimizations:

```
0.0007157325744628906 seconds: /preregistration/kiosk
0.0010983943939208984 seconds: /preregistration/paid_preregistrations
0.003024578094482422 seconds: /registration/new
0.004037141799926758 seconds: /accounts/insert_test_admin
0.005446434020996094 seconds: /magprime_dept_checklist/index
0.009880781173706055 seconds: /common/index
0.011798620223999023 seconds: /schedule/index
0.03175950050354004 seconds: /preregistration/index
0.03279256820678711 seconds: /tabletop_checkins/badged_attendees
0.037276506423950195 seconds: /accounts/reset
0.04175710678100586 seconds: /signups/jobs
0.047554731369018555 seconds: /registration/arbitrary_charge_form
0.047577857971191406 seconds: /registration/new_reg_station
0.04879164695739746 seconds: /registration/merch
0.048871517181396484 seconds: /preregistration/credit_card_retry
0.05038762092590332 seconds: /tabletop_tournaments/index
0.05077791213989258 seconds: /accounts/change_password
0.050829172134399414 seconds: /registration/attendee_upload
0.05448770523071289 seconds: /registration/discount
0.05514049530029297 seconds: /emails/emails_by_interest
0.05623769760131836 seconds: /accounts/homepage
0.05730271339416504 seconds: /tabletop_checkins/index
0.05845141410827637 seconds: /mivs_judging/index
0.060288429260253906 seconds: /graphs/index
0.06055307388305664 seconds: /signups/shirt_size
0.060628414154052734 seconds: /preregistration/check_prereg
0.06073760986328125 seconds: /summary/extra_merch
0.061721086502075195 seconds: /barcode/index
0.06336712837219238 seconds: /preregistration/stats
0.06477952003479004 seconds: /registration/review
0.06899404525756836 seconds: /budget/mpoints
0.06932234764099121 seconds: /registration/inactive
0.06955432891845703 seconds: /schedule/xml
0.07445859909057617 seconds: /signups/login
0.07519197463989258 seconds: /signups/shifts
0.08218669891357422 seconds: /devtools/index
0.08322334289550781 seconds: /preregistration/dealer_registration
0.08710455894470215 seconds: /schedule/schedule_tsv
0.08819794654846191 seconds: /tournament_admin/index
0.08951640129089355 seconds: /emails/pending
0.1054694652557373 seconds: /season_supporters/index
0.10785913467407227 seconds: /signups/printable
0.1088714599609375 seconds: /dept_checklist/index
0.1162407398223877 seconds: /magprime_dept_checklist/treasury
0.13111162185668945 seconds: /signups/index
0.13816189765930176 seconds: /registration/placeholders
0.14271235466003418 seconds: /magprime_dept_checklist/tech_requirements
0.14394235610961914 seconds: /static_views/magfest_js
0.14501237869262695 seconds: /schedule/now
0.15171575546264648 seconds: /schedule/panelists_owed_refunds
0.1585235595703125 seconds: /schedule/edit
0.18514561653137207 seconds: /emails/index
0.24964165687561035 seconds: /summary/affiliates
0.26171422004699707 seconds: /hotel/index
0.2847270965576172 seconds: /jobs/index
0.2864110469818115 seconds: /registration/index
0.2877359390258789 seconds: /jobs/signups
0.3762950897216797 seconds: /registration/feed
0.3827371597290039 seconds: /jobs/staffers
0.3909907341003418 seconds: /registration/recent
0.4291112422943115 seconds: /registration/comments
0.4394826889038086 seconds: /devtools/gitinfo
0.5932097434997559 seconds: /panel_app_management/index
0.6052346229553223 seconds: /accounts/sitemap
0.7396335601806641 seconds: /mivs_admin/index
0.7790632247924805 seconds: /hotel/hours
0.8397657871246338 seconds: /hotel_summary/setup_teardown
0.8541443347930908 seconds: /schedule/panels_json
0.8564319610595703 seconds: /summary/setup_teardown_neglect
0.8890223503112793 seconds: /jobs/summary
0.9166648387908936 seconds: /accounts/index
0.9332582950592041 seconds: /summary/found_how
0.9489150047302246 seconds: /jobs/everywhere
1.0246152877807617 seconds: /dept_checklist/overview
1.1168584823608398 seconds: /groups/index
1.2453687191009521 seconds: /registration/staffers
1.4690406322479248 seconds: /jobs/add_volunteers_by_dept
1.5005788803100586 seconds: /summary/ratings
1.5920054912567139 seconds: /jobs/all_shifts
1.6797709465026855 seconds: /summary/staffing_overview
1.7182462215423584 seconds: /summary/restricted_untaken
1.7558252811431885 seconds: /summary/departments
1.832895040512085 seconds: /summary/food_restrictions
1.8477790355682373 seconds: /hotel/requests
1.9761416912078857 seconds: /budget/index
1.9782240390777588 seconds: /summary/index
2.091369390487671 seconds: /hotel_assignments/index
2.2441470623016357 seconds: /summary/volunteers_owed_refunds
2.414466142654419 seconds: /hotel/no_shows
2.586413621902466 seconds: /summary/food_eligible
2.74497127532959 seconds: /summary/shirt_counts
2.846724510192871 seconds: /band_admin/index
5.002503395080566 seconds: /summary/consecutive_threshold
14.204671382904053 seconds: /summary/all_schedules
```